requirements.txt
# Core dependencies
google-generativeai>=0.3.0
google-adk>=0.1.0
python-dotenv>=1.0.0

# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Data handling
pydantic>=2.0.0
pydantic-settings>=2.0.0

# PDF parsing (for grant documents)
PyPDF2>=3.0.0
pdfplumber>=0.10.0

# Date/time handling
python-dateutil>=2.8.0

# Logging and monitoring
structlog>=23.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Deployment
gunicorn>=21.2.0
flask>=3.0.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0
